[Decomp][Spell] Cohesion split: extract SpellAuraDummy handlers to SpellAuraDummy.cpp
